feat: Added Longest Palindromic Subsequence
Finds the longest palindromic subsequence within the inputtes string. This can be the result of dropping letters from the string in order to acheive the longest palindrome.

E.g.  will return  because b can be dropped to allow a longer palindrome. If contiguous characters are required, look at Longest Palindromic Substring.

O(n) - which is the best this can be due to a need to evaluate every character for comparison.