fix: Added logic to deletion for BST
When the deletion logic must find the next highest value (that is less than the deleted value), there can be a conflict if the value for substitution has a rightChild.

Changed logic cleans this up, but applying the value of substitute.rightChild into it's parent's leftChild. This is okay, logically, because the substitute value will be from the left chain.